indexed_table: insert: Pass new value reference to eq closure
This can be required when using this with non-Copy values.

Hashbrown itself doesn't have an `insert` function and requires
splitting this into two steps using the entry API, where the caller
still holds the value when calling `entry` with the `eq` closure. The
second parameter is only needed when the caller already passed along
ownership of the value to test equality for.