table_seq: Begin review of `unsafe` usage
This adds a crate-wide `deny(unsafe_op_in_unsafe_fn)` together with the
necessary additional unsafe blocks this requires.

I also started adding `// SAFETY:` comments, after reviewing and
documenting the use of `unsafe` to eventually be able to add a
`#![warn(clippy::undocumented_unsafe_blocks)]` lint to ensure that every
use of unsafe comes with a comment explaining why it is safe.