ir: Fix a polarity bug when a rewrite changes the output polarity
The trait method for applying such a rewrite to a dynamically typed node
was losing the output polarity provided by the statically typed trait
method providing the actual implementation. This fixes this and also
adds a `#[must_use]` attribute to the statically typed trait method to
automatically flag such mistakes in the future.