ir: Rename Var's as_pos/as_neg to as_lit/as_neg_lit to avoid confusion
The Lit type has `as_pos` and `as_neg` types that force the polarity
ignoring the current polarity, while when giving a Var the current
polarity is known. Having the methods of both types being called the
same turns into a pitfall when calling either method on what is believed
to be a variable but actually is a literal. Having distinct names avoids
this.