inc_refine: Fix various corner cases and make generally more robust
This change adds timestamps to invalidate skip links, which is much less
error prone than the pervious strategy of ensuring that any invalid skip
link can be detected by where it is pointing to.

This also extends the provided API in that it adds more ways to query
the partition tree structure.